some errorchecking for existence of directories

+++ elevation/create-elevation-data-bg.bash
@@ -1,12 +1,38 @@
 #!/bin/bash
 
+# Reusable call of phyghtmap
+# --------------------------
 function createcontour {
   
+  # Get the given arguments
   ELESTEP=$1
-  ELECAT=$2
-  ELEPATH=$3
-  LOGFILE="logs/$MAPNAME-ele$ELESTEP.log"
+  ELEMEDIUM=$2
+  ELEMAJOR=$3
+
+  # Calculate some values used lateron
+  ELECAT="${ELEMAJOR},${ELEMEDIUM}"
+  ELEPATH="ele_${ELESTEP}_${ELEMEDIUM}_${ELEMAJOR}"
+  LOGFILE="logs/$MAPNAME-ele${ELESTEP}.log"
+
+  # Create the log directory if needed
+  if [ ! -d "logs" ]
+  then
+    mkdir logs
+  fi
   
+  # Create the elevation categories subdirectory if needed
+  if [ ! -d "pbf/${ELEPATH}" ]
+  then
+    mkdir -p "pbf/${ELEPATH}"
+  fi
+
+  # Creating some output
+  echo
+  echo "${ELESTEP}/${ELEMEDIUM}/${ELEMAJOR} m"
+  echo "---------------------------------------------"
+  
+  # Actually call phyghtmap
+  # -----------------------
   phyghtmap --step=${ELESTEP}                 \
             --osm-version=0.6                 \
             --jobs=1                          \
@@ -21,22 +47,29 @@ function createcontour {
             --output-prefix=./pbf/${ELEPATH}/Hoehendaten_${MAPNAME} | tee >(grep using | grep hgt | awk '{print $NF}' | sed 's/.$//' > $LOGFILE)
   
   
+  # if there is already a file with the needed name, remove it
   if [ -f ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}.osm.pbf ]
   then
      rm ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}.osm.pbf
   fi
-  mv ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}*.osm.pbf ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}.osm.pbf
+  
+  # Check if there is output created to be renamed
+  if [ -f ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}.osm.pbf ]
+  then
+    mv ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}*.osm.pbf ./pbf/${ELEPATH}/Hoehendaten_${MAPNAME}.osm.pbf
+  fi
   
 }
 
+# Main Part
+# ====================
+# Get the MAPNAME and set some defaults
 MAPNAME="$1"
 POLYFILE="./poly/$MAPNAME.poly"
-ELE10="ele_10_100_200"
-ELE20="ele_20_100_500"
-ELE25="ele_25_250_500"
 #DATASRC="view1,view3"
 DATASRC="view1,view3,srtm1,srtm3"
 
+# Check if ID's are given, else set default
 if [ -z "$2" ]
 then
    NID=7500000000
@@ -51,91 +84,13 @@ else
    WID=$3
 fi
 
+# Create the output
 echo 
 echo "$POLYFILE"
 echo "============================================="
 
-echo "20/100/500 m"
-echo "---------------------------------------------"
-createcontour "20" "500,100" "$ELE20"
+# Call the functions to really do the work
+createcontour 20 100 500
+createcontour 25 250 500
+createcontour 10 100 200
 
-echo
-echo "25/250/500 m"
-echo "---------------------------------------------"
-createcontour "25" "500,250" "$ELE20"
-
-echo "10/100/200 m"
-echo "---------------------------------------------"
-createcontour "10" "200,100" "$ELE20"
-
-
-#LOGFILE="logs/$MAPNAME-ele20.log"
-#phyghtmap --step=20                         \
-#          --osm-version=0.6                 \
-#          --jobs=1                          \
-#          --polygon=${POLYFILE}             \
-#          --line-cat=500,100                \
-#          --source=${DATASRC}               \
-#          --start-node-id=${NID}            \
-#          --start-way-id=${WID}             \
-#          --max-nodes-per-tile=0            \
-#          --pbf                             \
-#          --write-timestamp                 \
-#          --output-prefix=./pbf/${ELE20}/Hoehendaten_${MAPNAME} | tee >(grep using | grep hgt | awk '{print $NF}' | sed 's/.$//' > $LOGFILE)
-#
-#
-#if [ -f ./pbf/${ELE20}/Hoehendaten_${MAPNAME}.osm.pbf ]
-#then
-#   rm ./pbf/${ELE20}/Hoehendaten_${MAPNAME}.osm.pbf
-#fi
-#mv ./pbf/${ELE20}/Hoehendaten_${MAPNAME}*.osm.pbf ./pbf/${ELE20}/Hoehendaten_${MAPNAME}.osm.pbf
-#
-#echo
-#echo "25/250/500 m"
-#echo "---------------------------------------------"
-#LOGFILE="logs/$MAPNAME-ele25.log"
-#phyghtmap --step=25                         \
-#          --osm-version=0.6                 \
-#          --jobs=1                          \
-#          --polygon=${POLYFILE}             \
-#          --line-cat=500,250                \
-#          --source=${DATASRC}               \
-#          --start-node-id=${NID}            \
-#          --start-way-id=${WID}             \
-#          --max-nodes-per-tile=0            \
-#          --pbf                             \
-#          --write-timestamp                 \
-#          --output-prefix=./pbf/${ELE25}/Hoehendaten_${MAPNAME} | tee >(grep using | grep hgt | awk '{print $NF}' | sed 's/.$//' > $LOGFILE)
-#
-#if [ -f ./pbf/${ELE25}/Hoehendaten_${MAPNAME}.osm.pbf ]
-#then
-#   rm ./pbf/${ELE25}/Hoehendaten_${MAPNAME}.osm.pbf
-#fi
-#mv ./pbf/${ELE25}/Hoehendaten_${MAPNAME}*.osm.pbf ./pbf/$ELE25/Hoehendaten_${MAPNAME}.osm.pbf
-#
-#
-#echo 
-#echo "10/100/200 m"
-#echo "---------------------------------------------"
-#LOGFILE="logs/$MAPNAME-ele10.log"
-#phyghtmap --step=10                         \
-#          --osm-version=0.6                 \
-#          --jobs=1                          \
-#          --polygon=$POLYFILE               \
-#          --line-cat=200,100                \
-#          --source=${DATASRC}               \
-#          --start-node-id=${NID}            \
-#          --start-way-id=${WID}             \
-#          --max-nodes-per-tile=0            \
-#          --pbf                             \
-#          --write-timestamp                 \
-#          --output-prefix=./pbf/${ELE10}/Hoehendaten_${MAPNAME} | tee >(grep using | grep hgt | awk '{print $NF}' | sed 's/.$//' > $LOGFILE)
-#		  
-#if [ -f ./pbf/${ELE10}/Hoehendaten_${MAPNAME}.osm.pbf ]
-#then
-#   rm ./pbf/${ELE10}/Hoehendaten_${MAPNAME}.osm.pbf
-#fi
-#mv ./pbf/${ELE10}/Hoehendaten_${MAPNAME}*.osm.pbf ./pbf/${ELE10}/Hoehendaten_${MAPNAME}.osm.pbf
-#
-#echo
-#
